requirements.txt
fastapi
uvicorn
librosa
numpy
pandas
scikit-learn
openai-whisper
python-multipart
